visualizations/genea_visualizer/celery-queue/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
astunparse==1.6.3
attrs==19.3.0
Babel==2.9.1
billiard==3.5.0.3
bvh==0.3
celery==4.2.1
certifi==2020.6.20
chardet==3.0.4
cryptography==3.3.2
EasyProcess==0.3
flower==0.9.2
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.0
MarkupSafe==1.1.1
Parsley==1.3
piglet==1.0.0
piglet-templates==1.0.0
pycrypto==2.6.1
pygobject==3.26.1
pytz==2018.3
PyVirtualDisplay==1.3.2
pyxdg==0.26
redis==2.10.6
requests==2.24.0
SecretStorage==2.3.1
six==1.11.0
tornado==5.0.2
urllib3==1.26.5
vine==1.1.4
ffmpeg-python==0.2.0